Switch tests to parametrized py.test
The previous test suite used a weird "dynamic patching" of a UnitTest
class, where in fact all it needed was a sort of parametrization.

PLEASE NOTE: Some tests have been moved into an /extra folder, they're
testing some special cli flag that we're not really concerned with. The
aim is to include them at some point, but the priority is with the base
html->md tests.