Fix handling of emphasis
In HTML we have plenty of examples where an emphasis tag contains one or
more <br/> tags. The markdown specification isn't concerned with mapping
html to markdown, so we have to play the assume game. For our purposes,
and to my "common" sense, there's never a case when we want the closing
`**` to be on a different line.